fix(tokens): avoid filters in hooks
