+++ src/tokens/hooks.ts
@@ -1,5 +1,4 @@
 import BigNumber from 'bignumber.js'
-import DeviceInfo from 'react-native-device-info'
 import { TIME_UNTIL_TOKEN_INFO_BECOMES_STALE, TOKEN_MIN_AMOUNT } from 'src/config'
 import { usdToLocalCurrencyRateSelector } from 'src/localCurrency/selectors'
 import useSelector from 'src/redux/useSelector'
@@ -7,12 +6,13 @@ import { getDynamicConfigParams } from 'src/statsig'
 import { DynamicConfigs } from 'src/statsig/constants'
 import { StatsigDynamicConfigs } from 'src/statsig/types'
 import {
-  tokenCompareByUsdBalanceThenByName,
+  swappableTokensSelectorByNetworkId,
   tokensByAddressSelector,
   tokensByCurrencySelector,
   tokensByIdSelector,
   tokensListSelector,
   tokensListWithAddressSelector,
+  tokensWithTokenBalanceSelector,
   tokensWithUsdValueSelector,
   totalTokenBalanceSelector,
 } from 'src/tokens/selectors'
@@ -23,11 +23,9 @@ import {
   getSupportedNetworkIdsForSend,
   getSupportedNetworkIdsForTokenBalances,
   isCicoToken,
-  usdBalance,
 } from 'src/tokens/utils'
 import { NetworkId } from 'src/transactions/types'
 import { Currency } from 'src/utils/currencies'
-import { isVersionBelowMinimum } from 'src/utils/versionCheck'
 import networkConfig from 'src/web3/networkConfig'
 
 /**
@@ -49,44 +47,12 @@ export function useTotalTokenBalance() {
 
 export function useTokensWithTokenBalance() {
   const supportedNetworkIds = getSupportedNetworkIdsForTokenBalances()
-  const tokens = useSelector((state) => tokensListSelector(state, supportedNetworkIds))
-  return tokens.filter((tokenInfo) => tokenInfo.balance.gt(TOKEN_MIN_AMOUNT))
+  return useSelector((state) => tokensWithTokenBalanceSelector(state, supportedNetworkIds))
 }
 
 export function useTokensForSend() {
   const supportedNetworkIds = getSupportedNetworkIdsForSend()
-  const tokens = useSelector((state) => tokensListSelector(state, supportedNetworkIds))
-  return tokens.filter((tokenInfo) => tokenInfo.balance.gt(TOKEN_MIN_AMOUNT))
-}
-
-export function useTokensForAssetsScreen() {
-  const supportedNetworkIds = getSupportedNetworkIdsForTokenBalances()
-  const tokens = useSelector((state) => tokensListSelector(state, supportedNetworkIds))
-
-  return tokens
-    .filter((tokenInfo) => tokenInfo.balance.gt(TOKEN_MIN_AMOUNT) || tokenInfo.showZeroBalance)
-    .sort((token1, token2) => {
-      // Sorts by usd balance, then token balance, then zero balance natives by
-      // network id, then zero balance non natives by network id
-      const usdBalanceCompare = usdBalance(token2).comparedTo(usdBalance(token1))
-      if (usdBalanceCompare) {
-        return usdBalanceCompare
-      }
-
-      const balanceCompare = token2.balance.comparedTo(token1.balance)
-      if (balanceCompare) {
-        return balanceCompare
-      }
-
-      if (token1.isNative && !token2.isNative) {
-        return -1
-      }
-      if (!token1.isNative && token2.isNative) {
-        return 1
-      }
-
-      return token1.networkId.localeCompare(token2.networkId)
-    })
+  return useSelector((state) => tokensWithTokenBalanceSelector(state, supportedNetworkIds))
 }
 
 export function useTokensInfoUnavailable(networkIds: NetworkId[]) {
@@ -117,28 +83,11 @@ export function useTokenPricesAreStale(networkIds: NetworkId[]) {
   }
 }
 
-export function useSendableTokens() {
-  const networkIdsForSend = getDynamicConfigParams(
-    DynamicConfigs[StatsigDynamicConfigs.MULTI_CHAIN_FEATURES]
-  ).showSend
-  const tokens = useSelector((state) => tokensListSelector(state, networkIdsForSend))
-  return tokens.filter((tokenInfo) => tokenInfo.balance.gt(TOKEN_MIN_AMOUNT))
-}
-
 export function useSwappableTokens() {
-  const appVersion = DeviceInfo.getVersion()
   const networkIdsForSwap = getDynamicConfigParams(
     DynamicConfigs[StatsigDynamicConfigs.MULTI_CHAIN_FEATURES]
   ).showSwap
-  const tokens = useSelector((state) => tokensListSelector(state, networkIdsForSwap))
-  return tokens
-    .filter(
-      (tokenInfo) =>
-        tokenInfo.isSwappable ||
-        (tokenInfo.minimumAppVersionToSwap &&
-          !isVersionBelowMinimum(appVersion, tokenInfo.minimumAppVersionToSwap))
-    )
-    .sort(tokenCompareByUsdBalanceThenByName)
+  return useSelector((state) => swappableTokensSelectorByNetworkId(state, networkIdsForSwap))
 }
 
 export function useCashInTokens() {
